(fix) Address uwsgi segmentation fault issue
With Ubuntu bionic base image for drydock docker image, uwsig crashes
with segmentation fault when it tries to load/import the psycopg2 package.
The reason for this is that uwsgi and psycopg2 packages are built with
incopatible ssl libraries.

Upgrading uwsgi and psycopg2 to address this issue for bionic based
images.

Change-Id: I3d0bfb96c19849f5c2925366f8712bf47985df67